task_management_project/requirements.txt
asgiref==3.2.10
attrs==20.2.0
Django==3.1.2
djangorestframework==3.12.1
importlib-metadata==2.0.0
iniconfig==1.1.1
packaging==20.4
pika==1.1.0
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.1
pytest-django==4.1.0
pytest-mock==3.3.1
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
zipp==3.4.0
